requirements.txt
# 빌드 도구
setuptools>=65.5.1
wheel>=0.38.4

# 기본 의존성
numpy==1.21.6
scipy>=1.7.0
pillow==9.0.0

# PyTorch 관련
torch==2.1.0
torchvision==0.16.0
opencv-python-headless>=4.5.0
tqdm>=4.62.0
yapf>=0.40.0

# Flask 관련
werkzeug==2.0.1
flask==2.0.1
flask-cors==3.0.10

# 서버
gunicorn==20.1.0 